fahrenheit to celsius calculator using floats

+++ cpp/problems/clang/syntax005.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 main() {
-    int fahr, celsius;
-    int lower, upper, step;
+    float fahr, celsius;
+    float lower, upper, step;
 
     lower = 0;
     upper = 300;
@@ -10,8 +10,8 @@ main() {
 
     fahr = lower;
     while (fahr <= upper) {
-        celsius = 5 * (fahr - 32) / 9;
-        printf("%3d %6d\n", fahr, celsius);
+        celsius = 5.0 * (fahr - 32.0) / 9.0;
+        printf("%3.0f %6.1f\n", fahr, celsius);
         fahr = fahr + step;
     }
 }
